AMD64 changes:   * support for memory instrumentation   * improved instruction relocation (i.e. handling relocating     instructions outside a +/- 2GB range of data they're trying     to address)   * stackwalk fixes   * fixes for function relocation (thanks Drew & Nate)   * fix to get paradyn working (on bubba)     - note: this _was_ working, but has been broken since a more recent       CVS update.     - also note: even when this was working, it required the removal of       cap_stripped_binaries, since we were identifying something       incorrectly as a function which caused a segfault in our dominator       code

+++ dyninstAPI/h/BPatch_memoryAccess_NP.h
@@ -39,7 +39,7 @@
  * incur to third parties resulting from your use of Paradyn.
  */
 
-/* $Id: BPatch_memoryAccess_NP.h,v 1.17 2005/12/01 00:56:24 jaw Exp $ */
+/* $Id: BPatch_memoryAccess_NP.h,v 1.18 2005/12/12 16:37:05 gquinn Exp $ */
 
 #ifndef _MemoryAccess_h_
 #define _MemoryAccess_h_
@@ -73,16 +73,16 @@ extern void initOpCodeInfo();
 class BPATCH_DLL_EXPORT BPatch_addrSpec_NP
 {
   // the formula is regs[0] + 2 ^ scale * regs[1] + imm
-  int imm;      // immediate
+  long imm;      // immediate
   unsigned int scale;
   int regs[2];  // registers: -1 means none, 0 is 1st, 1 is 2nd and so on
                 // some pseudoregisters may be used, see the documentation
 
 public:
-  BPatch_addrSpec_NP(int _imm, int _ra = -1, int _rb = -1, int _scale = 0);
+  BPatch_addrSpec_NP(long _imm, int _ra = -1, int _rb = -1, int _scale = 0);
   BPatch_addrSpec_NP();
 
-  int getImm() const;
+  long getImm() const;
   int getScale() const;
   int getReg(unsigned i) const;
 
@@ -114,14 +114,14 @@ class BPATCH_DLL_EXPORT BPatch_memoryAccess : public BPatch_instruction
 
   // initializes only the first access - general case
   void set1st(bool _isLoad, bool _isStore,
-              int _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
-              int _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
+              long _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
+              long _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
               int _preFcn, int _cond, bool _nt);
 
   // initializes only the first access - no scale for count
   void set1st(bool _isLoad, bool _isStore,
-              int _imm_s, int _ra_s, int _rb_s,
-              int _imm_c, int _ra_c = -1, int _rb_c = -1,
+              long _imm_s, int _ra_s, int _rb_s,
+              long _imm_c, int _ra_c = -1, int _rb_c = -1,
               unsigned int _scale_s = 0, int _preFcn = -1,
               int _cond = -1, bool _nt = false);
 
@@ -132,51 +132,51 @@ class BPATCH_DLL_EXPORT BPatch_memoryAccess : public BPatch_instruction
   // initializes only the first access; #bytes is a constant
   BPatch_memoryAccess(const void *buf, int _sz, Address _addr,
 		      bool _isLoad, bool _isStore, unsigned int _bytes,
-            int _imm, int _ra, int _rb, unsigned int _scale = 0,
-            int _cond = -1, bool _nt = false);
+		      long _imm, int _ra, int _rb, unsigned int _scale = 0,
+		      int _cond = -1, bool _nt = false);
 
   // initializes only the first access; #bytes is an expression w/scale
   BPatch_memoryAccess(const void *buf, int _sz, Address _addr,
                       bool _isLoad, bool _isStore,
-                      int _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
-                      int _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
+                      long _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
+                      long _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
                       int _cond, bool _nt, int _preFcn = -1);
 
   // initializes only the first access; #bytes is an expression
   BPatch_memoryAccess(const void *buf, int _sz, Address _addr,
 		      bool _isLoad, bool _isStore, bool _isPrefetch,
-            int _imm_s, int _ra_s, int _rb_s,
-            int _imm_c, int _ra_c, int _rb_c,
-            unsigned short _preFcn);
+		      long _imm_s, int _ra_s, int _rb_s,
+		      long _imm_c, int _ra_c, int _rb_c,
+		      unsigned short _preFcn);
 
   // initializes only the first access; #bytes is an expression & not a prefetch
   BPatch_memoryAccess(const void *buf, int _sz, Address _addr,
-	       bool _isLoad, bool _isStore, int _imm_s, int _ra_s, int _rb_s,
-          int _imm_c, int _ra_c, int _rb_c);
+		      bool _isLoad, bool _isStore, long _imm_s, int _ra_s, int _rb_s,
+		      long _imm_c, int _ra_c, int _rb_c);
 
   // sets 2nd access; #bytes is constant
   void set2nd(bool _isLoad, bool _isStore, unsigned int _bytes,
-              int _imm, int _ra, int _rb, unsigned int _scale = 0);
+              long _imm, int _ra, int _rb, unsigned int _scale = 0);
 
   // sets 2nd access; #bytes is an expression w/scale
   void set2nd(bool _isLoad, bool _isStore,
-              int _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
-              int _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
+              long _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
+              long _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
               int _cond, bool _nt);
 
   // initializes both accesses; #bytes is a constant
   BPatch_memoryAccess(const void *buf, int _sz, Address _addr,
                       bool _isLoad, bool _isStore, unsigned int _bytes,
-                      int _imm, int _ra, int _rb, unsigned int _scale,
+                      long _imm, int _ra, int _rb, unsigned int _scale,
                       bool _isLoad2, bool _isStore2, unsigned int _bytes2,
-                      int _imm2, int _ra2, int _rb2, unsigned int _scale2);
+                      long _imm2, int _ra2, int _rb2, unsigned int _scale2);
 
   // initializes both accesses; #bytes is an expression & not a prefetch
   BPatch_memoryAccess(const void *buf, int _sz, Address _addr, bool _isLoad, bool _isStore,
-                      int _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
-                      int _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
-                      bool _isLoad2, bool _isStore2, int _imm2_s, int _ra2_s, 
-                      int _rb2_s, unsigned int _scale2_s, int _imm2_c, 
+                      long _imm_s, int _ra_s, int _rb_s, unsigned int _scale_s,
+                      long _imm_c, int _ra_c, int _rb_c, unsigned int _scale_c,
+                      bool _isLoad2, bool _isStore2, long _imm2_s, int _ra2_s, 
+                      int _rb2_s, unsigned int _scale2_s, long _imm2_c, 
                       int _ra2_c, int _rb2_c, unsigned int _scale2_c);
 
   virtual ~BPatch_memoryAccess();
